requirements.txt
streamlit==0.88.0
Pillow==8.3.2
py_avataaars==1.1.2
streamlit_lottie==0.0.2
